Refactoring and self review
