+++ packages/commonwealth/client/scripts/controllers/app/recent_activity.ts
@@ -3,7 +3,6 @@ import { Topic, AbridgedThread, Profile, Thread } from 'models';
 import app from 'state';
 import $ from 'jquery';
 import { modelFromServer as modelThreadFromServer } from 'controllers/server/threads';
-import ChainEntityController from 'controllers/server/chain_entities';
 
 export interface IAbridgedThreadFromServer {
   id: number;
