+++ packages/commonwealth/client/scripts/views/pages/discussions/thread_preview.tsx
@@ -12,7 +12,7 @@ import {
   getProposalUrlPath,
 } from 'identifiers';
 import { slugify } from 'utils';
-import { isCommandClick } from 'helpers';
+import { isCommandClick, pluralize } from 'helpers';
 import { AddressInfo, Thread } from 'models';
 import { ThreadPreviewReactionButton } from '../../components/reaction_button/thread_preview_reaction_button';
 import User from '../../components/widgets/user';
@@ -33,7 +33,6 @@ import {
   isWindowSmallInclusive,
 } from '../../components/component_kit/helpers';
 import { ThreadReactionButton } from '../../components/reaction_button/thread_reaction_button';
-// import { ThreadPreviewMenu } from './thread_preview_menu';
 
 type ThreadPreviewAttrs = {
   thread: Thread;
@@ -64,28 +63,9 @@ export class ThreadPreview extends ClassComponent<ThreadPreviewAttrs> {
   view(vnode: m.Vnode<ThreadPreviewAttrs>) {
     const { thread } = vnode.attrs;
 
-    // const commentsCount = app.comments.nComments(thread);
-
     const isSubscribed =
       getCommentSubscription(thread)?.isActive &&
       getReactionSubscription(thread)?.isActive;
-
-    // const hasAdminPermissions =
-    //   app.user.activeAccount &&
-    //   (app.roles.isRoleOfCommunity({
-    //     role: 'admin',
-    //     chain: app.activeChainId(),
-    //   }) ||
-    //     app.roles.isRoleOfCommunity({
-    //       role: 'moderator',
-    //       chain: app.activeChainId(),
-    //     }));
-
-    // const isAuthor =
-    //   app.user.activeAccount &&
-    //   thread.author === app.user.activeAccount.address;
-
-    // const canSeeMenu = hasAdminPermissions || isAuthor;
 
     return (
       <div
@@ -158,7 +138,7 @@ export class ThreadPreview extends ClassComponent<ThreadPreviewAttrs> {
           </div>
           <div class="title-row">
             <CWText type="h5" fontWeight="semiBold">
-              {(thread as unknown as any).numberOfComments}: {thread.title}
+              {thread.title}
             </CWText>
             {thread.hasPoll && <CWTag label="Poll" type="poll" />}
 
@@ -199,11 +179,10 @@ export class ThreadPreview extends ClassComponent<ThreadPreviewAttrs> {
               {this.isWindowSmallInclusive && (
                 <ThreadReactionButton thread={thread} />
               )}
-              {/* TODO Gabe 12/7/22 - Comment count isn't available before the comments store is initialized */}
-              {/* <CWIcon iconName="feedback" iconSize="small" />
+              <CWIcon iconName="feedback" iconSize="small" />
               <CWText type="caption">
-                {commentsCount} {!this.isWindowSmallInclusive && `comments`}
-              </CWText> */}
+                {pluralize(thread.numberOfComments, 'comment')}
+              </CWText>
             </div>
             <div class="row-bottom-menu">
               <div
