+++ packages/commonwealth/client/scripts/views/pages/overview/topic_summary_row.tsx
@@ -138,7 +138,6 @@ export class TopicSummaryRow extends ClassComponent<TopicSummaryRowAttrs> {
 
                   <div class="row-bottom">
                     <div class="comments-and-users">
-                      {/* TODO Gabe 12/7/22 - Comment count isn't available before the comments store is initialized */}
                       <div class="comments-count">
                         <CWIcon iconName="feedback" iconSize="small" />
                         <CWText type="caption">
